// src/app/interfaces/car.ts
export class Car {
    Id:number=0;
    CarImg: string="";
    CarType: string="";
    Make: string="";
    Model: string="";
    Milleage: string="";
    Fuel: string="";
    EngineSize: string="";
    MotorPower: string="";
    Gearbox: string="";
    SeatNumber: string="";
    Doors: string="";
    Color:string="";

}